Seperate crypto/string util functions from queue logic, Correct startup instructions in README
