+++ src/utils/string.utils.ts
@@ -1,0 +1,11 @@
+export class StringUtils {
+    public static  getGUID() {
+        var d = new Date().getTime();
+        var uuid: string = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+            var r = (d + Math.random() * 16) % 16 | 0;
+            d = Math.floor(d / 16);
+            return (c == 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+        });
+        return uuid;
+      };
+}
